requirements.txt
django
djangorestframework
deeppavlov
django-regex-field